Globally: Any and all instances of linking text should display as PMS 349 green on initial view. CMO: That's the color specified in the CSS. (more or less -- PMS colors are for INK, so you have to estimate, similar to how you would do to convert a PMS color to CMYK. Also, just like that conversion, the conversion for PMSC (Coated stock) and PMSU (Uncoated stock) is somewhat different. When presented with PMS colors (which usually I am not) I will generally go with the "coated" option as coated stock seems more similar to a screen presentation.
Are there specific instances where this is not the case? I don’t see any in the content areas of the site. Obviously, linked text in the footer cannot be green, it won’t be legible.

Home Page:
Learn More, Get Involved and Donate need to shifted right to be aligned with the right edge of the donate now! box. See New Top Bar Navigation[2].pdf DROPBOX