The top and bottom sliders on the Home Page behave differently. The top one disappears at a certain breakpoint, while the bottom one never does. The top one has an unvarying height, while the bottom one scales to a very large height. We actually prefer what is happening on the bottom slider.
CMO: The bottom “slider” is actually a custom navigation device, not a slideshow in the truest sense. It is trivial to leave the top slider on on all device sizes and make it responsive to the device width, however, making the top slider work like the How To Help navigation device would be a huge scope change at this point.